requirements.txt
paho-mqtt==2.0.0
pyzmq==25.1.2
cbor2==5.6.2